Move CA version to own package
