+++ cluster-autoscaler/version/version.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package main
+package version
 
 // ClusterAutoscalerVersion contains version of CA.
 const ClusterAutoscalerVersion = "1.14.0-beta.2"
